fix amountIndices size logic error

+++ Engine/src/Renderer.cpp
@@ -58,7 +58,7 @@ namespace basilisk
         glBufferData(GL_ARRAY_BUFFER, amountVertices * sizeof(float), vertices, GL_STATIC_DRAW);
         
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, this->Ebo);
-        glBufferData(GL_ELEMENT_ARRAY_BUFFER, amountIndices * sizeof(float), indices, GL_STATIC_DRAW);
+        glBufferData(GL_ELEMENT_ARRAY_BUFFER, amountIndices * sizeof(unsigned int), indices, GL_STATIC_DRAW);
         
         glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(float), static_cast<void*>(nullptr));
         glEnableVertexAttribArray(0);
